Try splitting tests up
