+++ website_portal_contract/tests/__init__.py
@@ -5,4 +5,5 @@
 from . import test_account_analytic_account
 from . import test_account_analytic_contract_template
 from . import test_account_analytic_contract
-from . import test_controller
+from . import test_controller_template
+from . import test_ui
