+++ website_portal_contract/tests/test_controller_template.py
@@ -25,13 +25,3 @@ class TestController(HttpCase):
             response.getcode(),
             200,
         )
-
-    def test_portal_contract_view_tour(self):
-        """ Tests contract view is correct """
-        tour = "odoo.__DEBUG__.services['web_tour.tour']"
-        self.phantom_js(
-            url_path='/my/home',
-            code="%s.run('test_contract_view')" % tour,
-            ready="%s.tours.test_contract_view.ready" % tour,
-            login='portal',
-        )
